requirements.txt
ax_platform==0.3.3
botorch==0.8.5
detector==3.0
gpytorch==1.10
hyperopt==0.2.7
matplotlib==3.6.2
numpy==1.23.4
pandas==1.5.1
PyYAML==6.0.1
scikit_learn==1.1.3
statsmodels==0.14.0
torch==2.0.1
tqdm==4.65.0
